requirements.txt
PyMuPDF>=1.22.0
Pillow>=10.0.0
easyocr>=1.7.1
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.19.0